# output-node/api/swagger.yaml
openapi: 3.0.2
info:
  title: DEV API
  description: |-
    Access DEV articles, comments and other resources via API

    # Authentication

    <!-- ReDoc-Inject: <security-definitions> -->
  termsOfService: https://dev.to/terms
  contact:
    name: DEV Team
    url: https://dev.to/contact
    email: yo@dev.to
  version: 0.5.7
  x-logo:
    url: https://practicaldev-herokuapp-com.freetls.fastly.net/assets/devlogo-pwa-128-1a9016906487b8ba17f2d3d3c28f9d1cb3ae9b45d685cc2711e20f6b3c6730df.png
    backgroundColor: '#000'
    altText: DEV logo
servers:
- url: https://dev.to/api
  description: Production server
paths:
...
  /webhooks:
    get:
      tags:
      - webhooks
      summary: Webhooks
      description: |
        This endpoint allows the client to retrieve a list of webhooks they have
        previously registered.

        "Webhooks" are used to register HTTP endpoints that will be called once a relevant event is
        triggered inside the web application, events like `article_created`, `article_updated`.

        It will return all webhooks, without pagination.
      operationId: webhooksGET
      responses:
        200:
          description: A list of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookIndex'
                x-content-type: application/json
              examples:
                webhooks-success:
                  summary: Successful response
                  value:
                  - type_of: webhook_endpoint
                    id: 1
                    source: DEV
                    target_url: https://example.com/webhooks/webhook1
                    events:
                    - article_created
                    created_at: 2019-09-02T09:47:39.230Z
      security:
      - api_key: []
      - oauth2: []
      x-code-samples:
      - lang: Shell
        label: curl
        source: |
          curl -H "api-key: API_KEY" https://dev.to/api/webhooks
      x-swagger-router-controller: Webhooks
    post:
      tags:
      - webhooks
      summary: Create a new webhook
      description: |
        This endpoint allows the client to create a new webhook.

        "Webhooks" are used to register HTTP endpoints that will be called once a relevant event is
        triggered inside the web application, events like `article_created`, `article_updated`.
      operationId: webhooksPOST
      requestBody:
        description: Webhook to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
            examples:
              webhook-create:
                $ref: '#/components/examples/WebhookCreate'
      responses:
        201:
          description: A newly created webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookShow'
              examples:
                webhook-success:
                  $ref: '#/components/examples/WebhookShow'
        400:
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                error-bad-request:
                  $ref: '#/components/examples/ErrorBadRequest'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                error-unauthorized:
                  $ref: '#/components/examples/ErrorUnauthorized'
        422:
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                error-unprocessable-entity:
                  $ref: '#/components/examples/ErrorUnprocessableEntity'
      security:
      - api_key: []
      - oauth2: []
      x-code-samples:
      - lang: Shell
        label: curl
        source: |
          curl -X POST -H "Content-Type: application/json" \
            -H "api-key: API_KEY" \
            -d '{"webhook_endpoint":{"target_url":"https://example.org/webhooks/webhook1","source":"DEV","events":["article_created"]}}' \
            https://dev.to/api/webhooks
      x-swagger-router-controller: Webhooks
  /webhooks/{id}:
    get:
      tags:
      - webhooks
      summary: A webhook endpoint
      description: |
        This endpoint allows the client to retrieve a single
        webhook given its `id`.
      operationId: webhooksIdGET
      parameters:
      - name: id
        in: path
        description: Id of the webhook
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
        example: 123
      responses:
        200:
          description: A webhook endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookShow'
              examples:
                webhook-success:
                  $ref: '#/components/examples/WebhookShow'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                article-unauthorized:
                  $ref: '#/components/examples/ErrorUnauthorized'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                article-not-found:
                  $ref: '#/components/examples/ErrorNotFound'
      security:
      - api_key: []
      - oauth2: []
      x-code-samples:
      - lang: Shell
        label: curl
        source: |
          curl https://dev.to/api/webhooks/123
      x-swagger-router-controller: Webhooks
    delete:
      tags:
      - webhooks
      summary: A webhook endpoint
      description: |
        This endpoint allows the client to delete a single
        webhook given its `id`.
      operationId: webhooksIdDELETE
      parameters:
      - name: id
        in: path
        description: Id of the webhook
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
        example: 123
      responses:
        204:
          description: A successful deletion
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                article-unauthorized:
                  $ref: '#/components/examples/ErrorUnauthorized'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                article-not-found:
                  $ref: '#/components/examples/ErrorNotFound'
      security:
      - api_key: []
      - oauth2: []
      x-code-samples:
      - lang: Shell
        label: curl
        source: |
          curl -X DELETE \
            -H "api-key: API_KEY" \
            https://dev.to/api/webhooks/1
      x-swagger-router-controller: Webhooks
components:
  schemas:
    APIError:
      required:
      - error
      - status
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
          format: int32
...
    ArticleUser:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        twitter_username:
          type: string
        github_username:
          type: string
        website_url:
          type: string
          format: url
        profile_image:
          type: string
          description: Profile image (640x640)
        profile_image_90:
          type: string
          description: Profile image (90x90)
      description: The article's creator
      example:
        profile_image: profile_image
        website_url: http://example.com/aeiou
        name: name
        twitter_username: twitter_username
        github_username: github_username
        profile_image_90: profile_image_90
        username: username
...
    WebhookCreate:
      type: object
      properties:
        webhook_endpoint:
          $ref: '#/components/schemas/WebhookCreate_webhook_endpoint'
      description: Webhook creation payload
    WebhookShow:
      type: object
      properties:
        type_of:
          type: string
        id:
          type: integer
          format: int64
        source:
          type: string
          description: The name of the requester, eg. "DEV"
        target_url:
          type: string
          format: url
        events:
          type: array
          description: An array of events identifiers
          items:
            type: string
        created_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/ArticleUser'
      description: Webhook
      example:
        target_url: http://example.com/aeiou
        created_at: 2000-01-23T04:56:07.000+00:00
        type_of: type_of
        id: 0
        source: source
        user:
          profile_image: profile_image
          website_url: http://example.com/aeiou
          name: name
          twitter_username: twitter_username
          github_username: github_username
          profile_image_90: profile_image_90
          username: username
        events:
        - events
        - events
    WebhookIndex:
      type: object
      properties:
        type_of:
          type: string
        id:
          type: integer
          format: int64
        source:
          type: string
          description: The name of the requester, eg. "DEV"
        target_url:
          type: string
          format: url
        events:
          type: array
          description: An array of events identifiers
          items:
            type: string
        created_at:
          type: string
          format: date-time
      description: Webhook
      example:
        target_url: http://example.com/aeiou
        created_at: 2000-01-23T04:56:07.000+00:00
        type_of: type_of
        id: 0
        source: source
        events:
        - events
        - events
...
    WebhookCreate_webhook_endpoint:
      required:
      - events
      - source
      - target_url
      type: object
      properties:
        source:
          type: string
          description: The name of the requester, eg. "DEV"
        target_url:
          type: string
          format: url
        events:
          type: array
          description: An array of events identifiers
          items:
            type: string
  examples:
    ErrorBadRequest:
      value:
        error: bad request
        status: 400
    ErrorNotFound:
      value:
        error: not found
        status: 404
    ErrorUnauthorized:
      value:
        error: unauthorized
        status: 401
    ErrorUnprocessableEntity:
      value:
        error: 'param is missing or the value is empty: article'
        status: 422
...
    WebhookCreate:
      value:
        webhook_endpoint:
          target_url: https://example.com/webhooks/webhook1
          source: DEV
          events:
          - article_created
    WebhookShow:
      value:
        type_of: webhook_endpoint
        id: 1
        source: DEV
        target_url: https://example.com/webhooks/webhook1
        events:
        - article_created
        created_at: 2019-09-02T09:47:39.230Z
        user:
          name: bob
          username: bob
          twitter_username: null
          github_username: bob
          website_url: null
          profile_image: '...'
          profile_image_90: '...'
  securitySchemes:
    api_key:
      type: apiKey
      description: |
        API Key authentication.

        Authentication for some endpoints, like write operations on the
        Articles API require a DEV API key.

        ### Getting an API key

        To obtain one, please follow these steps:

          - visit https://dev.to/settings/account
          - in the "DEV API Keys" section create a new key by adding a
            description and clicking on "Generate API Key"

            ![obtain a DEV API Key](https://user-images.githubusercontent.com/146201/64421366-af3f8b00-d0a1-11e9-8ff6-7cc0ca6e854e.png)
          - You'll see the newly generated key in the same view
            ![generated DEV API Key](https://user-images.githubusercontent.com/146201/64421367-af3f8b00-d0a1-11e9-9831-73d3bdfdff66.png)
      name: api-key
      in: header
    oauth2:
      type: oauth2
      description: |
        OAuth2 authentication.

        OAuth2 authentication is still in private alpha.
      flows:
        clientCredentials:
          tokenUrl: https://dev.to/oauth/token
          refreshUrl: https://dev.to/oauth/token
          scopes: {}
        authorizationCode:
          authorizationUrl: https://dev.to/oauth/authorize
          tokenUrl: https://dev.to/oauth/token
          refreshUrl: https://dev.to/oauth/token
          scopes: {}